+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Method to subscribe to register value updates.
+- Method to subscribe to emergency messages.
 
 ### Fixed
 - Feedback symmetry check calculation.